split cart display into separate component to load into /cart and /checkout
