+++ client/components/index.js
@@ -10,5 +10,6 @@ export {default as ProductsDisplay} from './multiple-product-display'
 export {default as SingleProductDisplay} from './single-product-display'
 export {default as ProductPage} from './product-page'
 export {default as Cart} from './Cart'
+export {default as CartDisplay} from './cart-display'
 export {default as Admin} from './Admin'
 export {default as Checkout} from './Checkout'
